rename: components 구조에 폴더를 하나 두어, style과 component들을 구분할 수 있도록 변경
