+++ src/features/common/components/header/PageHeader.tsx
@@ -2,7 +2,7 @@ import Image from 'next/image';
 import JiaryLogo from '@/static/jiary-logo.svg';
 import GithubLogo from '@/static/github-logo.svg';
 import UserProfile from '@/features/common/components/profile/UserProfile';
-import * as style from '@/features/common/components/header/PageHeader.css';
+import * as style from '@/features/common/components/header/PageHeader.css.ts';
 import Link from 'next/link';
 import { useSelector } from 'react-redux';
 import { isLoggedInSelector } from '@/store/slices/authSlice.ts';
