+++ src/features/diaryList/components/DiaryCardDropdown/DiaryCardDropdown.tsx
@@ -3,7 +3,7 @@ import Image from 'next/image';
 import { useRef, useState } from 'react';
 import useDeleteDoc from '@/features/diaryList/apis/mutations/useDeleteDoc.ts';
 import Dropdown from '@/features/common/components/dropdown/Dropdown.tsx';
-import * as style from '@/features/diaryList/components/DiaryCard.css';
+import * as style from '@/features/diaryList/components/DiaryCardDropdown/DiaryCardDropdown.css.ts';
 
 export default function DiaryCardDropdown({ id }: { id: string }) {
   const [isDropdownOpen, setIsDropdownOpen] = useState(false);
